Added input validation and more cuisine options
In this commit I added input validation to both of the user input fields, this way the user will continue to be prompted for an input until their input is one of the given options. I also added two more cuisine options to the program: Chinese restaurants and Indian Restaurants. I will continue to add more cuisine types and more QOL features.